Fixed issue with indentation (tabs vs spaces!)
Used:
python -m tabnanny <filename.py>
to fix the lines where there was tabs that should have been spaces.
Also updated the version number to 2.1.0 and status In Production.
This version also adds the access log function used to put an entry in a
new table in Seltzer that keeps track of who came in when.